Properly handle both regularly saved files and files automatically saved on hot exit
Due to hot exit events not currently made availble by ST, it's necessary to capture close events.

This causes a problem on dirty views, as it's not possible to distinguish a user-initiated close from a hot exit:

- in the former case, we don't want to save the fold data
- in the latter case, we do

There is no workaround possible using ST events; this solution stores the folding data on a per-view content basis (using a checksum), so that on open, the plugin will pick the fold data corresponding to the view content loaded.

On close events, previous fold data versions are kept. On save, only the latest one is kept, since it's guaranteed that on open, the saved version of the file will be opened.

If a storage file for a previous version is found, it's reset.